+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/transaction/ExecuteOnDataSourceBeanPostProcessor.java
@@ -11,6 +11,7 @@ import org.springframework.aop.support.AopUtils;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.core.annotation.AnnotationConfigurationException;
+import ru.hh.nab.jdbc.annotation.ExecuteOnDataSource;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jdbc.routing.DataSourceContextUnsafe;
