+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/transaction/ExecuteOnDataSourceTransactionCallback.java
@@ -7,6 +7,8 @@ import org.aspectj.lang.ProceedingJoinPoint;
 import org.springframework.lang.NonNull;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.support.TransactionCallback;
+import ru.hh.nab.jdbc.annotation.DataSourceCacheMode;
+import ru.hh.nab.jdbc.annotation.ExecuteOnDataSource;
 
 public class ExecuteOnDataSourceTransactionCallback implements TransactionCallback<Object> {
