+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/transaction/ExecuteOnDataSourceTransactionCallbackFactory.java
@@ -2,6 +2,7 @@ package ru.hh.nab.hibernate.transaction;
 
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.springframework.transaction.support.TransactionCallback;
+import ru.hh.nab.jdbc.annotation.ExecuteOnDataSource;
 
 public interface ExecuteOnDataSourceTransactionCallbackFactory {
   TransactionCallback<Object> create(ProceedingJoinPoint pjp, ExecuteOnDataSource executeOnDataSource);
