+++ nab-jdbc/src/main/java/ru/hh/nab/jdbc/annotation/ExecuteOnDataSource.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate.transaction;
+package ru.hh.nab.jdbc.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
