+++ nab-tests/src/test/java/ru/hh/nab/datasource/ExecuteOnDataSourceAspectTest.java
@@ -26,10 +26,10 @@ import org.springframework.test.context.ContextConfiguration;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isActualTransactionActive;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isSynchronizationActive;
 import ru.hh.nab.common.properties.FileSettings;
-import ru.hh.nab.hibernate.transaction.DataSourceCacheMode;
-import ru.hh.nab.hibernate.transaction.ExecuteOnDataSource;
 import ru.hh.nab.hibernate.transaction.ExecuteOnDataSourceAspect;
 import ru.hh.nab.hibernate.transaction.ExecuteOnDataSourceTransactionCallbackFactory;
+import ru.hh.nab.jdbc.annotation.DataSourceCacheMode;
+import ru.hh.nab.jdbc.annotation.ExecuteOnDataSource;
 import static ru.hh.nab.jdbc.common.DataSourceType.MASTER;
 import static ru.hh.nab.jdbc.common.DataSourceType.READONLY;
 import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.getDataSourceName;
